pam_unix: turn daysleft from int to long
The daysleft value could be larger than INT_MAX depending on system
configuration (content of shadow file).

We already cover this case with a subtract function, but it's much
easier to simply turn daysleft into a long. This simplifies the code
and is even a no-op on 32 bit platforms (if subtract is optimized
away).

+++ modules/pam_unix/passverify.h
@@ -73,7 +73,7 @@ PAMH_ARG_DECL(int get_pwd_hash,
 	const char *name, struct passwd **pwd, char **hash);
 
 PAMH_ARG_DECL(int check_shadow_expiry,
-	struct spwd *spent, int *daysleft);
+	struct spwd *spent, long *daysleft);
 
 PAMH_ARG_DECL(int unix_update_passwd,
 	const char *forwho, const char *towhat);
